% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{get_filename_cgc}
\alias{get_filename_cgc}
\title{Get URL of CGC strain list}
\usage{
get_filename_cgc()
}
\value{
a vector of length 2 giving the URL path and the filename.
}
\description{
Get URL of CGC strain list
}
