% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{get_filename_transcriptome}
\alias{get_filename_transcriptome}
\title{Get FTP path and filename of the transcriptome FASTA file}
\usage{
get_filename_transcriptome(WS, protocol = "https")
}
\arguments{
\item{WS}{Wormbase release number.}

\item{protocol}{use https or ftp}
}
\value{
a vector of length 2 giving the FTP path and the filename. Note the filename changes between versions!
}
\description{
This internal function gets the Wormbase FTP location and the filename
of the transcriptome FASTA file for a given Wormbase release.
}
