% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{get_filename_txdb}
\alias{get_filename_txdb}
\title{Get filename for a TxDb file}
\usage{
get_filename_txdb(WS)
}
\arguments{
\item{WS}{Wormbase release number.}
}
\value{
a vector of length 1 with the filename.
}
\description{
This internal function sets the name to use to store a TxDb file in cache.
}
