% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_WS}
\alias{get_WS}
\title{Validate Wormbase release number.}
\usage{
get_WS(WS)
}
\arguments{
\item{WS}{Release number to validate.}
}
\value{
A number labelling a valid Wormbase release.
}
\description{
This internal function enables other functions to receive the Wormbase release
number as a numerical value or a character.
}
