% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_WS_from_date.R
\name{guess_WS_from_date}
\alias{guess_WS_from_date}
\title{Guess Wormbase release from date}
\usage{
guess_WS_from_date(date)
}
\arguments{
\item{date}{a Date object or an object that can be coerced into one.}
}
\value{
An estimate of the current production release of Wormbase at that date.
}
\description{
Note these estimates are approximatives, for exact dates please refer to the official
release schedule available at: [https://wormbase.org/about/release_schedule]
}
\examples{
guess_WS_from_date("2018-01-01")
guess_WS_from_date("2028-04-10")
guess_WS_from_date("2008-04-01")
}
