% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2i.R
\name{i2s}
\alias{i2s}
\title{Convert Wormbase ID to symbol}
\usage{
i2s(gene_id, geneIDs)
}
\arguments{
\item{gene_id}{Wormbase gene IDs}

\item{geneIDs}{Translation table generated by `load_gene_ids()`.}
}
\value{
The corresponding gene symbols. If a gene symbol is missing, the sequence
name is returned instead.
}
\description{
Convert Wormbase ID to symbol
}
\examples{
gids <- wb_load_gene_ids("WS277")
s2i("unc-10", gids)
}
