% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{if_missing}
\alias{if_missing}
\title{Set the Value of a Missing Function Argument}
\usage{
if_missing(x, true = NA, false = x)
}
\arguments{
\item{x}{A function argument}

\item{true}{What to return if x is missing. Default is `NA`}

\item{false}{What to return if x is not missing. Default to return itself}
}
\description{
A simple wrapper around ifelse with the condition set to missing(x)
}
