% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gene_ids.R
\name{load_gene_ids}
\alias{load_gene_ids}
\title{Load Wormbase Gene ID table}
\usage{
load_gene_ids(WS, dir_cache = "~/../Projects/ref")
}
\arguments{
\item{WS}{Wormbase release version (e.g. WS277).}

\item{dir_cache}{Directory where the downloaded files are cached.}
}
\value{
A tibble obtained from reading the Wormbase table.
}
\description{
Load a cached geneIDs table. This gives the correspondence between gene symbol (e.g. unc-10),
Wormbase ID (e.g. WBGene00006750) and the sequence name (e.g. T10A3.1).
}
\examples{
gids <- load_gene_ids("WS277")
s2i("unc-10", gids)
}
