% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{refresh_cache_on_load}
\alias{refresh_cache_on_load}
\title{Set wbdata cache to automatically refresh on loading wbdata}
\usage{
refresh_cache_on_load(refresh_on_load = TRUE, tim = 7,
  units = "days")
}
\arguments{
\item{units}{}
}
\description{
Set wbdata cache to automatically refresh on loading wbdata
}
