% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_gene_names.R
\name{wb_clean_gene_names}
\alias{wb_clean_gene_names}
\title{Clean up old gene names}
\usage{
wb_clean_gene_names(
  gene_id,
  warn_missing = TRUE,
  dir_cache = NULL,
  refresh = 20,
  return_one = FALSE
)
}
\arguments{
\item{gene_id}{Vector of gene IDs to clean.}

\item{warn_missing}{Warn if some gene IDs are not found.}

\item{dir_cache}{Directory where the downloaded files are cached.}

\item{refresh}{Number of days after which the downloaded list needs to be refreshed.}

\item{return_one}{if `FALSE`, a single input may return several gene IDs separated by commas when the gene was split. If set to `TRUE`, only the first match is returned.}
}
\value{
An updated list of gene IDs.
}
\description{
Takes a list of (potentially old) Wormbase gene IDs, and replaces the Dead ones
with current IDs. This corresponds to the [Gene Name Sanitizer](https://wormbase.org/tools/mine/gene_sanitizer.cgi)
available on Wormbase.

Note this function only works with gene IDs (of the form WBGene00000001), not
symbols or other sequence names, while the online tool can potentially correct
old gene names. The file downloaded by the (unexported) function `wb_get_gene_name_history()`
contains additional information, ask if you need this functionality to be added to the package.
}
\examples{
\dontrun{
gids <- wb_load_gene_ids(294)
genes_of_interest <- c("WBGene00012733", "WBGene00000424", "WBGene00007566", "WBGene00045409")
i2s(genes_of_interest, gids, warn_missing = TRUE)
cleaned_genes <- wb_clean_gene_names(genes_of_interest)
i2s(cleaned_genes, gids, warn_missing = TRUE)
}
}
