% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_gene_names.R
\name{wb_get_gene_name_history}
\alias{wb_get_gene_name_history}
\title{Load the gene name history}
\usage{
wb_get_gene_name_history(dir_cache = NULL, refresh = 20)
}
\arguments{
\item{dir_cache}{Directory where the downloaded files are cached.}

\item{refresh}{Number of days after which the downloaded list needs to be refreshed.}
}
\value{
A tibble with "Input" containing historical gene names, "Suggested Match" the suggested match, and other fields
}
\description{
Load the gene name history
}
