% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gtf.R
\name{wb_get_gtf_path}
\alias{wb_get_gtf_path}
\title{Get path to cached GTF file}
\usage{
wb_get_gtf_path(WS, dir_cache = NULL)
}
\arguments{
\item{WS}{Wormbase release version.}

\item{dir_cache}{Directory where the downloaded files are cached.}
}
\value{
a character string pointing to a "c_elegans.PRJNA13758.WSxxx.canonical_geneset.gtf.gz" file.
}
\description{
Returns the path to a GTF file in the cache. Useful to use with functions that require
a connection to read from (e.g. to create a TxDb for Bioconductor packages). The file gets
downloaded as necessary.
}
\details{
For the TxDB creation, see also `wb_load_txdb()` which does it automatically.
}
\examples{
\dontrun{
library(GenomicFeatures)
library(wbData)

path_to_gtf <- wb_get_gtf_path(273)
my_txdb <- makeTxDbFromGFF(path_to_gtf,dataSource = "Wormbase",organism = "Caenorhabditis elegans")
gids <- wb_load_gene_ids(273)
select(my_txdb, keys = s2i("nduo-3", gids), columns = "TXSTART", keytype = "GENEID")
}
}
