% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cgc.R
\name{wb_load_cgc_list}
\alias{wb_load_cgc_list}
\title{Load the list of strains available at the CGC}
\usage{
wb_load_cgc_list(dir_cache = NULL, refresh = 2)
}
\arguments{
\item{dir_cache}{Directory where the downloaded files are cached.}

\item{refresh}{Number of days after which the downloaded list needs to be refreshed.}
}
\value{
A tibble containing the list of strains available at the CGC. Description of the fields can be found on the CGC website.
}
\description{
Load the list of strains available at the CGC
}
\examples{
strain_list <- wb_load_cgc_list()
nrow(strain_list)
strain_list$Genotype[strain_list$Strain == "NC902"]
}
