% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gtf.R
\name{wb_load_exon_coords}
\alias{wb_load_exon_coords}
\title{Load exon coordinates}
\usage{
wb_load_exon_coords(WS, dir_cache = NULL)
}
\arguments{
\item{WS}{Wormbase release version.}

\item{dir_cache}{Directory where the downloaded files are cached.}
}
\value{
A tibble obtained from reading the Wormbase GTF, containing the
following fields:
\itemize{
  \item \code{exon_id}: Unique exon ID, composed of the transcript name
  with the exon number.
  \item \code{transcript_id}: Transcript ID, typically the sequence name,
  an isoform letter (if relevant) and a transcript number (if relevant).
  \item \code{gene_id}: Wormbase gene ID (WBGene).
  \item \code{exon_number}: An integer identifying exons of each transcript.
  \item \code{chr}: Chromosome name (I, II, III, IV, V, X, MtDNA).
  \item \code{start}, \code{end}, \code{strand}: Genomic position of the exon.
  \item \code{gene_biotype}: Gene biotype (e.g. protein_coding or tRNA).
  \item \code{transcript_biotype}: Trtanscript biotype, usually the same as the gene biotype.
  \item \code{position}: Genomic position (compatible with JBrowse or IGV).
}
}
\description{
Loads a data frame providing the genomic coordinates of all exons, built from
the Wormbase GTF file. The GTF file will be downloaded as needed.
}
\examples{
\dontrun{
exon_coords <- wb_load_exon_coords("WS277")
exon_coords[exon_coords$transcript_id == "K08H10.7.2", "position"]
}

}
