% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gtf.R
\name{wb_load_gene_coords}
\alias{wb_load_gene_coords}
\title{Load gene coordinates}
\usage{
wb_load_gene_coords(WS, dir_cache = NULL)
}
\arguments{
\item{WS}{Wormbase release version.}

\item{dir_cache}{Directory where the downloaded files are cached.}
}
\value{
A tibble obtained from reading the Wormbase GTF, containing the
following fields:
\itemize{
  \item \code{gene_id}: Wormbase gene ID (WBGene).
  \item \code{chr}: Chromosome name (I, II, III, IV, V, X, MtDNA).
  \item \code{start}, \code{end}, \code{strand}: Genomic position of the gene.
  \item \code{position}: Genomic position (compatible with JBrowse or IGV).
  \item \code{gene_biotype}: Gene biotype (e.g. protein_coding or tRNA)
}
}
\description{
Loads a data frame providing the genomic coordinates of all genes, built from
the Wormbase GTF file. The GTF file will be downloaded as needed.
}
\examples{
\dontrun{
  gene_coords <- wb_load_gene_coords("WS273")
  gene_coords[gene_coords$gene_id == "WBGene00004323", "position"]
}
}
