% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gtf.R
\name{wb_load_tx2gene}
\alias{wb_load_tx2gene}
\title{Load transcript to gene correspondence table}
\usage{
wb_load_tx2gene(WS, dir_cache = NULL)
}
\arguments{
\item{WS}{Wormbase release version.}

\item{dir_cache}{Directory where the downloaded files are cached.}
}
\value{
A tibble obtained from reading the Wormbase GTF, containing the following fields:
\itemize{
  \item \code{gene_id}: Wormbase gene ID (WBGene).
  \item \code{transcript_id}: A transcript identifier, e.g. F46H6.2b.1 or T05A10.6.1.
  \item \code{gene_biotype}: Gene biotype (e.g. protein_coding or tRNA).
  \item \code{transcript_biotype}: Transcript biotype, usually the same as the gene biotype (see below).
}
}
\description{
Loads a data frame providing the transcripts and corresponding gene IDs, built from
the Wormbase GTF file. The GTF file will be downloaded as needed.
}
\section{Biotypes}{

The gene and transcrit biotypes are typically identical. However a few miRNA genes have their transcript annotated
as \code{pre_miRNA} or \code{miRNA_primary_transcript}, some \code{ncRNA}s are contained within
\code{protein_coding} genes, and there are a handful of tRNA and rRNA pseudogenes.
}

\examples{
\dontrun{
  tx2g_tab <- wb_load_tx2gene("WS281")
  table(tx2g_tab$gene_biotype, tx2g_tab$transcript_biotype)
  wb_tx2g("MTCE.10", tx2g_tab)
  wb_g2tx("WBGene00014457", tx2g_tab)
}
}
