% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_txdb.R
\name{wb_load_TxDb}
\alias{wb_load_TxDb}
\title{Load TxDb}
\usage{
wb_load_TxDb(WS, dir_cache = NULL)
}
\arguments{
\item{WS}{Wormbase release version.}

\item{dir_cache}{Directory where the downloaded files are cached.}
}
\value{
A TxDb object.
}
\description{
Generate and load a TxDb object for a given Wormbase version, that can be used with Bioconductor packages.
}
\examples{
\dontrun{
wb_txdb_274 <- wb_load_TxDb(274)
transcript_to_gene_table <- AnnotationDbi::select(wb_txdb_274,
                                         keys = keys(wb_txdb_274, keytype = "TXNAME"),
                                         columns = c("TXNAME", "GENEID"),
                                         keytype = "TXNAME")
all_exons_GRanges <- GenomicFeatures::exonsBy(wb_txdb_274, by = "gene")
}

}
