% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbdata-package.R
\docType{package}
\name{wbdata-package}
\alias{wbdata}
\alias{wbdata-package}
\title{wbdata: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jpiburn/wbdata}
  \item Report bugs at \url{https://github.com/jpiburn/wbdata/issues}
}

}
\author{
\strong{Maintainer}: Jesse Piburn \email{jesse.piburn@gmail.com}

}
\keyword{internal}
