% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{wbData}
\alias{wbData}
\alias{wbData-package}
\title{World Bank data.}
\format{A data frame with 11342 rows and 12 variables:
\describe{
  \item{country}{country name, factor with 214 levels}
  \item{country.code}{iso-2 code, factor with 213 levels}
  \item{isoAlpha3}{isoAlpha3 code, factor with 212 levels}
  \item{continent}{continent name, factor with 6 levels}
  \item{year}{ranges from 1960 to 2012}
  \item{land.area}{land area, in square kilometers}
  \item{life.exp}{life expectancy at birth, in years}
  \item{fertility.rate}{the number of children that would be born to a woman}
  \item{population}{total population}
  \item{urban.pop}{percentage of urban population of total population}
  \item{above65.pop}{percentage of population aged 65 or above of total population}
  \item{gdp.per.capita}{GDP per capita in US dollars}
  }}
\source{
\url{http://data.worldbank.org/}
}
\description{
Excerpt of the World Bank indicator data on life expectancy, GDP per capita, population,
land area, urban population(% of total), population aged 65 and above(% total) and fertility rate
by country, every year from 1960 to 2012.
}
\examples{
if (require("dplyr")) {
wbData
wbData \%>\%
  filter(year == 2012) \%>\%
  group_by(continent) \%>\%
  summarise(life.exp = median(life.exp))
}
}

