% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gene_ids.R
\name{get_filename}
\alias{get_filename}
\title{Find the filename and FTP path}
\usage{
get_filename(type, WS)
}
\arguments{
\item{type}{Type of file to obtain, e.g. geneID or GTF.}

\item{WS}{Wormbase release version.}
}
\value{
A length 2 vector containing the FTP path to the file and the name of the file.
}
\description{
For now, only geneID is a recognized file type. If needed, GTF or other types
of data could be added here.
}
