% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2i.R
\name{s2i}
\alias{s2i}
\title{Convert symbol to Wormbase ID}
\usage{
s2i(symbol, geneIDs)
}
\arguments{
\item{symbol}{gene symbols.}

\item{geneIDs}{Translation table generated with `load_gene_ids()`.}
}
\value{
Wormbase ID.
}
\description{
Convert symbol to Wormbase ID
}
\examples{
gids <- load_gene_ids("WS277")
s2i("unc-10", gids)
}
